Add index step overrides to c2a
Previously, the indices of table entries would increase by 1 for each entry. Now, you can specify the increase amount in the first cell of the table.

For instance, a first cell with `.byte 0 1` is equivalent to the old `.byte 0` in which the processed table definitions would look like

```
foo = 0
bar = 1
baz = 2
...
```

A first cell of `.byte 0 2` will have definitions that look like
```
foo = 0
bar = 2
baz = 4
...
```